Refactor `mergeAllOf` and `dedupAllOfRefs` logic into `Refs` augmenter
