+++ tests/Augmenter/InheritanceTest.php
@@ -8,9 +8,11 @@ namespace OpenApi\Tests\Augmenter;
 
 use OpenApi\Assembler;
 use OpenApi\Augmenter;
+use OpenApi\Compiler\OpenApi31Compiler;
 use OpenApi\Tests\Concerns\AssertsSchemaStructure;
 use OpenApi\Tests\Fixtures\Augmenter\Hierarchy\Spec as Fixtures;
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\Yaml\Yaml;
 
 final class InheritanceTest extends TestCase
 {
@@ -33,6 +35,7 @@ final class InheritanceTest extends TestCase
 
         $spec = $assembler->getSpecification();
         (new Augmenter\Inheritance())($spec);
+        (new Augmenter\Refs())($spec);
 
         $this->assertSpecificationSchemasMatchFile(
             $spec,
